requirements.txt
py_cpuinfo==9.0.0
